fix(profile): move edit profile into settings
