+++ src/components/main/settings/pages/profile/styles.scss
@@ -1,0 +1,11 @@
+.status {
+    // flex: auto;
+}
+
+.input_status {
+    width: 60%;
+}
+.change-status {
+    display: flex;
+    justify-content: space-between;
+}
